\name{DOTermsAnnDbBimap}
\docType{class}

\alias{DOTermsAnnDbBimap}
\alias{class:DOTermsAnnDbBimap}
\alias{DOTermsAnnDbBimap-class}



% Standard generic methods

\title{Class "DOTermsAnnDbBimap" }

\description{
  A sub-class of  Bimap, specific for DO.db. Please see \code{\link[AnnotationDbi:Bimap]{Bimap}} for details.
}


\seealso{
   \code{\link[AnnotationDbi:Bimap]{Bimap}},
   \code{\link{DOTerms}}
}

\examples{
   class(DOTERM)
}

\keyword{classes}
\keyword{interface}
